+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# [1.3.0] – 2022-11-24
+
+Fully support DokuWiki release Igor.
+
+- Handle deprecated functions `html_edit` and `html_diff`
+
 # [1.2.0] – 2021-11-14
 
 Prepare for upcoming DokuWiki release Igor.